Isabela

+++ UdeAStay_Desafio2/anfitrion.h
@@ -6,7 +6,7 @@ class Anfitrion {
 private:
     char* documento;
     int antiguedad;
-    int puntuacion;
+    float puntuacion;
     int* codigosAlojamientos;
     int numAlojamientos;
